Renamed Providers to mirror MSAL.Net docs
