+++ Microsoft.Graph.Auth/Microsoft.Graph.Auth.Test/ConfidentialClient/AuthorizationCodeProviderTests.cs
@@ -1,20 +1,19 @@
 ﻿namespace Microsoft.Graph.Auth.Test.ConfidentialClient
 {
-    using System;
-    using System.Collections.Generic;
-    using System.Net.Http;
-    using System.Threading.Tasks;
     using Microsoft.Graph.Auth.Test.Mocks;
     using Microsoft.Identity.Client;
     using Microsoft.VisualStudio.TestTools.UnitTesting;
     using NSubstitute;
+    using System;
+    using System.Collections.Generic;
+    using System.Net.Http;
+    using System.Threading.Tasks;
 
     [TestClass]
-    public class BehalfOfFlowProviderTests
+    public class AuthorizationCodeProviderTests
     {
         private TokenCache tokenCache;
-        private UserAssertion userAssertion;
-        private BehalfOfFlowProvider authCodeFlowProvider;
+        private AuthorizationCodeProvider authCodeFlowProvider;
         private IConfidentialClientApplication confidentialClientAppMock;
         private const string clientId = "client-id";
         private const string commonAuthority = "https://login.microsoftonline.com/common/";
@@ -31,45 +30,35 @@
             mockUserAccount2 = new MockUserAccount("abc@test.com", "login.microsoftonline.com");
             confidentialClientAppMock = Substitute.For<IConfidentialClientApplication>();
             tokenCache = new TokenCache();
-            var mockAuthResult = MockAuthResult.GetAuthenticationResult(scopes);
-            userAssertion = new UserAssertion(mockAuthResult.AccessToken);
-            authCodeFlowProvider = new BehalfOfFlowProvider(clientId, redirectUri, new ClientCredential(appSecret), tokenCache, scopes, userAssertion);
+            authCodeFlowProvider = new AuthorizationCodeProvider(clientId, redirectUri, new ClientCredential(appSecret), tokenCache, scopes);
         }
 
         [TestMethod]
-        public void BehalfOfFlow_ConstructorWithoutAuthority()
+        public void AuthorizationCodeProvider_ConstructorWithoutAuthority()
         {
             Assert.IsInstanceOfType(authCodeFlowProvider, typeof(IAuthenticationProvider));
             Assert.IsNotNull(authCodeFlowProvider.ClientApplication);
             Assert.IsInstanceOfType(authCodeFlowProvider.ClientApplication, typeof(IConfidentialClientApplication));
-            Assert.AreEqual(authCodeFlowProvider.ClientApplication.ClientId, clientId);
-            Assert.AreEqual(authCodeFlowProvider.ClientApplication.RedirectUri, redirectUri);
-            Assert.AreEqual(authCodeFlowProvider.ClientApplication.Authority, commonAuthority);
         }
 
         [TestMethod]
-        public void BehalfOfFlow_ConstructorWithAuthority()
+        public void AuthorizationCodeProvider_ConstructorWithAuthority()
         {
-            var authProvider = new BehalfOfFlowProvider(clientId, organizationsAuthority, redirectUri, new ClientCredential(appSecret), tokenCache, scopes, userAssertion);
+            var authProvider = new AuthorizationCodeProvider(clientId, commonAuthority, redirectUri, new ClientCredential(appSecret), tokenCache, scopes);
 
             Assert.IsInstanceOfType(authProvider, typeof(IAuthenticationProvider));
             Assert.IsNotNull(authProvider.ClientApplication);
             Assert.IsInstanceOfType(authProvider.ClientApplication, typeof(IConfidentialClientApplication));
-            Assert.AreEqual(authProvider.ClientApplication.ClientId, clientId);
-            Assert.AreEqual(authProvider.ClientApplication.RedirectUri, redirectUri);
-            Assert.AreEqual(authProvider.ClientApplication.Authority, organizationsAuthority);
         }
 
         [TestMethod]
-        public async Task BehalfOfFlow_WithNoUserAccountInCache()
+        public async Task AuthorizationCodeProvider_WithNoUserAccountInCache()
         {
             var httpRequestMessage = new HttpRequestMessage(HttpMethod.Get, "http://example.org/foo");
             var expectedAuthResult = MockAuthResult.GetAuthenticationResult(scopes);
-            var mockAuthResult = MockAuthResult.GetAuthenticationResult(scopes);
-
+            
             confidentialClientAppMock.GetAccountsAsync().ReturnsForAnyArgs(new List<IAccount>());
-            confidentialClientAppMock.AcquireTokenSilentAsync(scopes, mockUserAccount).ReturnsForAnyArgs(mockAuthResult);
-            confidentialClientAppMock.AcquireTokenOnBehalfOfAsync(scopes, userAssertion, commonAuthority).ReturnsForAnyArgs(expectedAuthResult);
+            confidentialClientAppMock.AcquireTokenByAuthorizationCodeAsync(string.Empty, scopes).ReturnsForAnyArgs(expectedAuthResult);
 
             authCodeFlowProvider.ClientApplication = confidentialClientAppMock;
 
@@ -85,13 +74,11 @@
 
             Assert.IsInstanceOfType(authCodeFlowProvider, typeof(IAuthenticationProvider));
             Assert.IsInstanceOfType(authCodeFlowProvider.ClientApplication, typeof(IConfidentialClientApplication));
-            Assert.IsNotNull(httpRequestMessage.Headers.Authorization);
-            Assert.AreEqual(httpRequestMessage.Headers.Authorization.Scheme, CoreConstants.Headers.Bearer);
-            Assert.AreEqual(httpRequestMessage.Headers.Authorization.Parameter, expectedAuthResult.AccessToken);
+            Assert.IsNull(httpRequestMessage.Headers.Authorization);
         }
 
         [TestMethod]
-        public async Task BehalfOfFlow_WithUserAccountInCache()
+        public async Task AuthorizationCodeProvider_WithUserAccountInCache()
         {
             var httpRequestMessage = new HttpRequestMessage(HttpMethod.Get, "http://example.org/foo");
             var expectedAuthResult = MockAuthResult.GetAuthenticationResult(scopes);
@@ -99,7 +86,7 @@
 
             confidentialClientAppMock.GetAccountsAsync().ReturnsForAnyArgs(new List<IAccount> { mockUserAccount });
             confidentialClientAppMock.AcquireTokenSilentAsync(scopes, null).ReturnsForAnyArgs(expectedAuthResult);
-            confidentialClientAppMock.AcquireTokenOnBehalfOfAsync(scopes, userAssertion, commonAuthority).ReturnsForAnyArgs(mockAuthResult);
+            confidentialClientAppMock.AcquireTokenByAuthorizationCodeAsync(string.Empty, scopes).ReturnsForAnyArgs(mockAuthResult);
 
             authCodeFlowProvider.ClientApplication = confidentialClientAppMock;
