+++ Microsoft.Graph.Auth/Microsoft.Graph.Auth.Test/Microsoft.Graph.Auth.Test.csproj
@@ -84,18 +84,18 @@
     <Reference Include="System.Xml.Linq" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="ConfidentialClient\AuthenticationCodeFlowProviderTests.cs" />
-    <Compile Include="ConfidentialClient\BehalfOfFlowProviderTests.cs" />
-    <Compile Include="ConfidentialClient\ClientCredentialFlowProviderTests.cs" />
+    <Compile Include="ConfidentialClient\AuthorizationCodeProviderTests.cs" />
+    <Compile Include="ConfidentialClient\OnBehalfOfProviderTests.cs" />
+    <Compile Include="ConfidentialClient\ClientCredentialProviderTests.cs" />
     <Compile Include="Extensions\StringExtensions.cs" />
     <Compile Include="Mocks\MockAuthResult.cs" />
     <Compile Include="Mocks\MockUserAccount.cs" />
     <Compile Include="Mocks\StorageMock.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="PublicClient\DeviceCodeFlowProviderTests.cs" />
-    <Compile Include="PublicClient\InteractiveFlowProviderTests.cs" />
-    <Compile Include="PublicClient\IntergratedWindowsAuthFlowProviderTests.cs" />
-    <Compile Include="PublicClient\UsernamePasswordFlowProviderTests.cs" />
+    <Compile Include="PublicClient\DeviceCodeProviderTests.cs" />
+    <Compile Include="PublicClient\InteractiveAuthenticationProviderTests.cs" />
+    <Compile Include="PublicClient\IntergratedWindowsAuthenticationProviderTests.cs" />
+    <Compile Include="PublicClient\UsernamePasswordProviderTests.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
