+++ Microsoft.Graph.Auth/Microsoft.Graph.Auth.Test/PublicClient/IntergratedWindowsAuthenticationProviderTests.cs
@@ -12,10 +12,10 @@ using System.Threading.Tasks;
 namespace Microsoft.Graph.Auth.Test.PublicClient
 {
     [TestClass]
-    public class InteractiveFlowProviderTests
+    public class IntergratedWindowsAuthenticationProviderTests
     {
         private TokenCache tokenCache;
-        private InteractiveFlowProvider interactiveFlowProvider;
+        private IntergratedWindowsAuthenticationProvider intergratedWindowsAuthFlowProvider;
         private IPublicClientApplication publicClientAppMock;
         private string clientId = "client-id";
         private string commonAuthority = "https://login.microsoftonline.com/common/";
@@ -29,21 +29,21 @@ namespace Microsoft.Graph.Auth.Test.PublicClient
             mockUserAccount = new MockUserAccount("xyz@test.net", "login.microsoftonline.com");
             mockUserAccount2 = new MockUserAccount("abc@test.com", "login.microsoftonline.com");
             tokenCache = new TokenCache();
-            interactiveFlowProvider = new InteractiveFlowProvider(clientId, scopes);
+            intergratedWindowsAuthFlowProvider = new IntergratedWindowsAuthenticationProvider(clientId, organizationsAuthority, tokenCache, scopes);
         }
 
         [TestMethod]
-        public void InteractiveFlow_ConstructorWithNullableParams()
+        public void IntergratedWindowsAuthenticationProvider_ConstructorWithNullableParams()
         {
-            Assert.IsInstanceOfType(interactiveFlowProvider, typeof(IAuthenticationProvider));
-            Assert.IsNotNull(interactiveFlowProvider.ClientApplication);
-            Assert.IsInstanceOfType(interactiveFlowProvider.ClientApplication, typeof(IPublicClientApplication));
+            Assert.IsInstanceOfType(intergratedWindowsAuthFlowProvider, typeof(IAuthenticationProvider));
+            Assert.IsNotNull(intergratedWindowsAuthFlowProvider.ClientApplication);
+            Assert.IsInstanceOfType(intergratedWindowsAuthFlowProvider.ClientApplication, typeof(IPublicClientApplication));
         }
 
         [TestMethod]
-        public void InteractiveFlow_ConstructorWithoutNullableParams()
+        public void IntergratedWindowsAuthenticationProvider_ConstructorWithoutNullableParams()
         {
-            var authProvider = new InteractiveFlowProvider(clientId, scopes, mockUserAccount, commonAuthority, tokenCache, UIBehavior.SelectAccount, "extra", null, new UIParent());
+            var authProvider = new IntergratedWindowsAuthenticationProvider(clientId, organizationsAuthority, tokenCache, scopes, mockUserAccount.Username);
 
             Assert.IsInstanceOfType(authProvider, typeof(IAuthenticationProvider));
             Assert.IsNotNull(authProvider.ClientApplication);
@@ -51,44 +51,48 @@ namespace Microsoft.Graph.Auth.Test.PublicClient
         }
 
         [TestMethod]
-        public async Task InteractiveFlow_WithNoUserAccountInCache()
+        public async Task IntergratedWindowsAuthenticationProvider_WithNoUserAccountInCache()
         {
             var httpRequestMessage = new HttpRequestMessage(HttpMethod.Get, "http://example.org/foo");
             var expectedAuthResult = MockAuthResult.GetAuthenticationResult(scopes);
             var mockAuthResult = MockAuthResult.GetAuthenticationResult(scopes);
+            var mockAuthResult2 = MockAuthResult.GetAuthenticationResult(scopes);
 
             publicClientAppMock = Substitute.For<IPublicClientApplication>();
             publicClientAppMock.GetAccountsAsync().ReturnsForAnyArgs(new List<IAccount>());
-            publicClientAppMock.AcquireTokenAsync(scopes, string.Empty, UIBehavior.SelectAccount, null, null, null, null).ReturnsForAnyArgs(expectedAuthResult);
-            publicClientAppMock.AcquireTokenAsync(scopes, mockUserAccount, UIBehavior.SelectAccount, null, null, null, null).ReturnsForAnyArgs(mockAuthResult);
+            publicClientAppMock.AcquireTokenSilentAsync(scopes, mockUserAccount).ReturnsForAnyArgs(mockAuthResult);
+            publicClientAppMock.AcquireTokenByIntegratedWindowsAuthAsync(scopes, mockUserAccount.Username).ReturnsForAnyArgs(mockAuthResult2);
+            publicClientAppMock.AcquireTokenByIntegratedWindowsAuthAsync(scopes).ReturnsForAnyArgs(expectedAuthResult);
 
-            interactiveFlowProvider.ClientApplication = publicClientAppMock;
-            await interactiveFlowProvider.AuthenticateRequestAsync(httpRequestMessage);
+            intergratedWindowsAuthFlowProvider.ClientApplication = publicClientAppMock;
+            await intergratedWindowsAuthFlowProvider.AuthenticateRequestAsync(httpRequestMessage);
 
-            Assert.IsInstanceOfType(interactiveFlowProvider, typeof(IAuthenticationProvider));
-            Assert.IsInstanceOfType(interactiveFlowProvider.ClientApplication, typeof(IPublicClientApplication));
+            Assert.IsInstanceOfType(intergratedWindowsAuthFlowProvider, typeof(IAuthenticationProvider));
+            Assert.IsInstanceOfType(intergratedWindowsAuthFlowProvider.ClientApplication, typeof(IPublicClientApplication));
             Assert.IsNotNull(httpRequestMessage.Headers.Authorization);
             Assert.AreEqual(httpRequestMessage.Headers.Authorization.Scheme, CoreConstants.Headers.Bearer);
             Assert.AreEqual(httpRequestMessage.Headers.Authorization.Parameter, expectedAuthResult.AccessToken);
         }
 
         [TestMethod]
-        public async Task InteractiveFlow_WithUserAccountInCache()
+        public async Task IntergratedWindowsAuthenticationProvider_WithUserAccountInCache()
         {
             var httpRequestMessage = new HttpRequestMessage(HttpMethod.Post, "http://example.org/foo");
             var expectedAuthResult = MockAuthResult.GetAuthenticationResult(scopes);
             var mockAuthResult = MockAuthResult.GetAuthenticationResult(scopes);
+            var mockAuthResult2 = MockAuthResult.GetAuthenticationResult(scopes);
 
             publicClientAppMock = Substitute.For<IPublicClientApplication>();
             publicClientAppMock.GetAccountsAsync().ReturnsForAnyArgs(new List<IAccount> { mockUserAccount });
             publicClientAppMock.AcquireTokenSilentAsync(scopes, mockUserAccount).ReturnsForAnyArgs(expectedAuthResult);
-            publicClientAppMock.AcquireTokenAsync(scopes, mockUserAccount, UIBehavior.SelectAccount, null, null, null, null).Returns(mockAuthResult);
+            publicClientAppMock.AcquireTokenByIntegratedWindowsAuthAsync(scopes, mockUserAccount.Username).ReturnsForAnyArgs(mockAuthResult);
+            publicClientAppMock.AcquireTokenByIntegratedWindowsAuthAsync(scopes).ReturnsForAnyArgs(mockAuthResult2);
 
-            interactiveFlowProvider.ClientApplication = publicClientAppMock;
-            await interactiveFlowProvider.AuthenticateRequestAsync(httpRequestMessage);
+            intergratedWindowsAuthFlowProvider.ClientApplication = publicClientAppMock;
+            await intergratedWindowsAuthFlowProvider.AuthenticateRequestAsync(httpRequestMessage);
 
-            Assert.IsInstanceOfType(interactiveFlowProvider, typeof(IAuthenticationProvider));
-            Assert.IsInstanceOfType(interactiveFlowProvider.ClientApplication, typeof(IPublicClientApplication));
+            Assert.IsInstanceOfType(intergratedWindowsAuthFlowProvider, typeof(IAuthenticationProvider));
+            Assert.IsInstanceOfType(intergratedWindowsAuthFlowProvider.ClientApplication, typeof(IPublicClientApplication));
             Assert.IsNotNull(httpRequestMessage.Headers.Authorization);
             Assert.AreEqual(httpRequestMessage.Headers.Authorization.Scheme, CoreConstants.Headers.Bearer);
             Assert.AreEqual(httpRequestMessage.Headers.Authorization.Parameter, expectedAuthResult.AccessToken);
